Rmeoved SaveGame Class
Kenny phoned and we discussed serializing the Game object and killing
SaveGame. Done and done! Great idea, KC!

Also, I've changed the params for the Game constructor to be just
Captain and planets... the Planets class will have a flag for
exloration. Puzzles and bosses will be tied to planets (so visiting
those planets will be tantamount to beating them). 

Need: Captain and Persona classes!