CrewSelection and CrewConfirmation is Done!
I have updated the captain and gamecontroller2 classes to correctly
choose and confirm a captain's crew. Also, added a 3 crew member test to
the CaptainClassTester class. 

A note about the tester: it can only be used on a game that has asked
the captain for a crew. A test on a test-game will always fail.